+++ src/cli/commands/list.ts
@@ -3,9 +3,9 @@ import {
   ensureOpen,
   hasParameter,
   type KVDBContainer,
-  userInput,
 } from "../common.ts";
 import { KVKeyInstance, type KVQuery } from "../../lib/key.ts";
+import { printTransaction } from "../common.ts";
 
 export async function list(
   container: KVDBContainer,
@@ -16,20 +16,24 @@ export async function list(
 
   let query: KVQuery;
   if (hasParameter(params, 0)) {
-    query = KVKeyInstance.parse(params[0], true); // Query parsing
+    // Validate query
+    try {
+      const parsedKey = KVKeyInstance.parse(params[0], true);
+      query = new KVKeyInstance(parsedKey, true).get();
+    } catch (e) {
+      console.error(`Could not parse query: ${e.message}`);
+      return false;
+    }
   } else {
-    const queryInput = userInput("Enter query (dot separated): ");
-    if (!queryInput) return false; // Exit if no query provided
-    query = KVKeyInstance.parse(queryInput, true);
+    console.error("No query supplied.");
+    return false;
   }
 
   console.log("");
 
-  // Iterate over matching entries
-  for await (const entry of container.db!.iterate(query)) {
-    // Display key information
-    const key = new KVKeyInstance(entry.key).stringify();
-    console.log(key, entry.data);
+  // Iterate over matching entries (max 1000)
+  for await (const entry of container.db!.iterate(query, 1000)) {
+    printTransaction(entry);
   }
 
   console.log(""); // Extra newline for separation