+++ src/lib/kv.ts
@@ -9,7 +9,7 @@ import {
   type KVTransactionResult,
 } from "./transaction.ts";
 import { KVLedger } from "./ledger.ts";
-import { SYNC_INTERVAL_MS } from "./constants.ts";
+import { LEDGER_CACHE_MB, SYNC_INTERVAL_MS } from "./constants.ts";
 
 // External dependencies
 import { EventEmitter } from "node:events";
@@ -82,6 +82,28 @@ export interface KVOptions {
    * @defaultValue `1000`
    */
   syncIntervalMs?: number;
+
+  /**
+   * The maximum size (in megabytes) of raw ledger data to cache in memory.
+   *
+   * Note that actual memory usage may be slightly higher due to the associated overhead of storing index data and metadata.
+   *
+   * @defaultValue `10` (10 MB)
+   */
+  ledgerCacheSize?: number;
+
+  /**
+   * Disables the in-memory index, leading to faster loading,
+   * but preventing efficient querying and iteration using `get`, `iterate`, `scan`, and `list`.
+   *
+   * When `disableIndex` is true:
+   * - Only the `set`, `delete`, and `scan` operations are available.
+   * - Use cases are limited to scenarios where only appending to the ledger is required.
+   * - Memory usage is minimized as the in-memory index is not maintained.
+   *
+   * @defaultValue `false`
+   */
+  disableIndex?: boolean;
 }
 
 /**
@@ -99,10 +121,12 @@ export class KV extends EventEmitter {
 
   // Configuration
   private ledgerPath?: string;
+  private ledgerCacheSize: number = LEDGER_CACHE_MB;
   /** Public only for testing purposes */
   public autoSync: boolean = true;
   /** Public only for testing purposes */
   public syncIntervalMs: number = SYNC_INTERVAL_MS;
+  private disableIndex = false;
 
   // States
   private blockSync: boolean = false; // Syncing can be blocked during vacuum
@@ -139,6 +163,25 @@ export class KV extends EventEmitter {
       );
     }
     this.syncIntervalMs = options.syncIntervalMs ?? SYNC_INTERVAL_MS;
+    // - ledgerCacheSize
+    if (
+      options.ledgerCacheSize !== undefined &&
+      (!Number.isInteger(options.ledgerCacheSize) ||
+        options.ledgerCacheSize <= 0)
+    ) {
+      throw new TypeError(
+        "Invalid option: ledgerCacheSize must be a positive integer",
+      );
+    }
+    this.ledgerCacheSize = options.ledgerCacheSize ?? this.ledgerCacheSize;
+    // - disableIndex
+    if (
+      options.disableIndex !== undefined &&
+      typeof options.disableIndex !== "boolean"
+    ) {
+      throw new TypeError("Invalid option: disableIndex must be a boolean");
+    }
+    this.disableIndex = options.disableIndex ?? false;
 
     if (this.autoSync) {
       this.watchdogPromise = this.watchdog();
@@ -166,7 +209,7 @@ export class KV extends EventEmitter {
     }
 
     // Open the ledger, and start a new watchdog
-    this.ledger = new KVLedger(filePath);
+    this.ledger = new KVLedger(filePath, this.ledgerCacheSize);
     this.ledgerPath = filePath;
     await this.ledger.open(createIfMissing);
 
@@ -259,15 +302,14 @@ export class KV extends EventEmitter {
     try {
       if (doLock) await this.ledger?.lock();
 
-      const newTransactions = await this.ledger?.sync();
+      const newTransactions = await this.ledger?.sync(this.disableIndex);
 
       if (newTransactions === null) { // Ledger invalidated
         result = "ledgerInvalidated";
         await this.open(this.ledgerPath!, false); // Reopen ledger
       } else {
         result = newTransactions?.length ? "success" : "ready"; // Success if new transactions exist
-
-        if (newTransactions) {
+        if (newTransactions && !this.disableIndex) {
           for (const entry of newTransactions) {
             try {
               this.applyTransactionToIndex(entry.transaction, entry.offset); // Refactored for clarity
@@ -322,8 +364,8 @@ export class KV extends EventEmitter {
    * @throws {Error} If the database is not open or if the transaction operation is unsupported.
    */
   private applyTransactionToIndex(transaction: KVTransaction, offset: number) {
-    // Throw if database isn't open
     this.ensureOpen();
+    this.ensureIndex();
 
     // Check for matches in watch handlers
     for (const handler of this.watchHandlers) {
@@ -346,6 +388,19 @@ export class KV extends EventEmitter {
   }
 
   /**
+   * Function that throws of the index is disabled.
+   *
+   * @throws {Error} If the index is disabled.
+   */
+  private ensureIndex(): void {
+    if (this.disableIndex) {
+      throw new Error(
+        "Operation not available due to `disableIndex` option being set.",
+      );
+    }
+  }
+
+  /**
    * Performs a vacuum operation to reclaim space in the underlying ledger.
    *
    * This operation is essential for maintaining performance as the database grows over time.
@@ -360,6 +415,7 @@ export class KV extends EventEmitter {
   public async vacuum(): Promise<void> {
     // Throw if database isn't open
     this.ensureOpen();
+    this.ensureIndex();
 
     this.blockSync = true;
     await this.ledger?.vacuum();
@@ -413,7 +469,7 @@ export class KV extends EventEmitter {
   ): Promise<KVTransactionResult<T> | null> {
     // Throw if database isn't open
     this.ensureOpen();
-
+    this.ensureIndex();
     for await (const entry of this.iterate<T>(key, 1)) {
       return entry;
     }
@@ -448,7 +504,7 @@ export class KV extends EventEmitter {
   ): AsyncGenerator<KVTransactionResult<T>> {
     // Throw if database isn't open
     this.ensureOpen();
-
+    this.ensureIndex();
     const validatedKey = new KVKeyInstance(key, true);
     const offsets = this.index!.get(validatedKey, limit)!;
 
@@ -481,6 +537,7 @@ export class KV extends EventEmitter {
   ): Promise<KVTransactionResult<T>[]> {
     // Throw if database isn't open
     this.ensureOpen();
+    this.ensureIndex();
 
     const entries: KVTransactionResult<T>[] = [];
     for await (const entry of this.iterate<T>(key)) {
@@ -504,6 +561,7 @@ export class KV extends EventEmitter {
   public count(key: KVQuery): number {
     // Throw if database isn't open
     this.ensureOpen();
+    this.ensureIndex();
 
     const validatedKey = new KVKeyInstance(key, true);
     const offsets = this.index.get(validatedKey);
@@ -526,10 +584,8 @@ export class KV extends EventEmitter {
 
     // Ensure the key is ok
     const validatedKey = new KVKeyInstance(key);
-
     const transaction = new KVTransaction();
 
-    performance.mark("b");
     await transaction.create(
       validatedKey,
       KVOperation.SET,
@@ -635,12 +691,29 @@ export class KV extends EventEmitter {
       const baseOffset = await this.ledger!.add(transactionsData);
 
       // Update the index and check for errors
-      for (const { transaction, relativeOffset } of bufferedTransactions) {
+      for (
+        const { transaction, transactionData, relativeOffset }
+          of bufferedTransactions
+      ) {
         try {
-          this.applyTransactionToIndex(
-            transaction,
+          // Add to ledger cache
+          this.ledger!.cache.cacheTransactionData(
             baseOffset + relativeOffset,
+            {
+              offset: baseOffset + relativeOffset,
+              length: transactionData.length,
+              complete: true,
+              transaction,
+            },
           );
+
+          // Add to index
+          if (!this.disableIndex) {
+            this.applyTransactionToIndex(
+              transaction,
+              baseOffset + relativeOffset,
+            );
+          }
         } catch (error) {
           errors.push(error as Error);
         }
@@ -678,6 +751,7 @@ export class KV extends EventEmitter {
   public listKeys(key: KVKey | KVQuery | null): string[] {
     // Throw if database isn't open
     this.ensureOpen();
+    this.ensureIndex();
 
     return this.index.getChildKeys(
       key === null ? null : new KVKeyInstance(key, true),
@@ -695,6 +769,7 @@ export class KV extends EventEmitter {
     callback: (transaction: KVTransactionResult<T>) => void,
     recursive: boolean = false,
   ) {
+    this.ensureIndex();
     this.watchHandlers.push({ query, callback, recursive });
   }
 
@@ -712,6 +787,7 @@ export class KV extends EventEmitter {
     query: KVQuery,
     callback: (transaction: KVTransactionResult<T>) => void,
   ): boolean {
+    this.ensureIndex();
     const newWatchHandlers = this.watchHandlers.filter(
       (handler) => handler.query !== query || handler.callback !== callback,
     );